Add override for broken callback after timeout
Sencha's ScriptTagProxy calls the request's callback with a null operation
after a timeout. This differs from the usual ServerProxy error behavior
of calling the usual callback with a valid operation, and is separate
code because a JSONP script tag embed can only fail by timemout, so must
be actively checked.

Fix to call the operation's callback.