Fixes to main js file
Switch to ScriptTagProxy so we can do cross domain live loads
of schedule date. Load via a Yahoo! Pipe to get JSONP from the
PyGotham.org hosted JSON file.

Set flex on List for height layout.

New JSON uses a simple string for the level classification
(beginner/intermediate/advanced) if there is only one level
specified; test for this to avoid type errors calling .join()
when level is a string, not a list of strings.