Brutal hack: add TZ offset to naive timestamps
Safari (desktop and mobile), and thus the iPhone, chokes on ISO datetimes
without timezone offsets. The schedule JSON gives exactly such, and our
ISO8601 parser also fails to catch them, leading to falling back on the
browser's parsing.

Make the custom 8601 code a bit more forgiving (" " accepted as well as
"T" for time/date separator).

However, the piece that makes it actually work is a horrible, horrible
hack: Append "-04:00" for EDT (in effect for the time and location of the
conference – NYC on 8 June 2012) in the Sencha Touch Model convert function
before we pass the value to the ISO8601 code. Argh. But it works, and the
timezone's not changing for the conference at this point.